backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector database
faiss-cpu==1.7.4  # or faiss-gpu if you have GPU
numpy==1.26.2
pickle5  # for compatibility

# OpenAI
openai==1.6.0
tiktoken==0.5.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0

# Data processing
pandas==2.1.3
beautifulsoup4==4.12.0